moved cli interface into the ui module
