+++ raiden/main.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from ui.cli import run
+
+
+if __name__ == '__main__':
+    run()
