+++ raiden/ui/console.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from __future__ import print_function
+
 import cStringIO
 import json
 import sys
